requirements.txt
moviepy>=1.0.3
Pillow>=9.5.0
pydub>=0.25.1
srt>=3.5.3
requests>=2.31.0
google-api-python-client>=2.111.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
PyYAML>=6.0
pytrends>=4.9.0
numpy>=1.24.0
pandas>=1.5.0
